🎨 style: [#30] constants 성격에 맞게 파일 분리
