+++ src/constants/path.ts
@@ -1,0 +1,12 @@
+export const PAGE_URL = {
+  ROOT: '/',
+  NOT_FOUND: '*',
+  SIGN_IN: '/auth/signin',
+  SIGN_UP: '/auth/signup',
+  MYSTAT: '/mystat',
+  MYPAGE: '/mypage',
+  SEARCHBOOK: '/book/search',
+  POSTBOOK: '/book/register/:id',
+  DETAILBOOKINFO: '/book/detail/:id',
+  RECOMMENDEDBOOKS: '/book/recommended',
+};
