+++ src/constants/recommend.ts
@@ -1,0 +1,12 @@
+export const RECOMMENDED_SORT = [
+  {
+    id: 0,
+    name: '베스트셀러',
+    value: 'best-seller',
+  },
+  {
+    id: 1,
+    name: '주목할만한 신간 리스트',
+    value: 'item-new-special',
+  },
+];
